# openapi.yaml
openapi: 3.0.3
info:
  title: OpenStocks API
  version: 1.1.0
  description: |
    Tokenized pre-IPO trading platform (take-home).

    Stretch features included:
    - GBM simulated price ticks
    - WebSocket price streaming (`/prices` namespace, `price` event)
    - HTTP rate limiting (429 when exceeded; configure via `THROTTLE_*`)
    - Short selling with 50% initial margin
    - Price-time-priority matching across users
    - Admin halt/resume per symbol
servers:
  - url: http://localhost:3000
paths:
...
  /admin/prices/{symbol}:
    post:
      summary: Set simulated mark price and match the resting book (price-time priority)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [price]
              properties:
                price:
                  type: number
                ts:
                  type: string
                  format: date-time
      responses:
        "200":
          description: Updated
        "429":
          description: Rate limited
  /admin/halt/{symbol}:
    post:
      summary: Manually halt trading for a symbol (independent of circuit breaker)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Halted
        "400":
          description: Unknown symbol
        "401":
          description: Unauthorized
        "429":
          description: Rate limited
  /admin/resume/{symbol}:
    post:
      summary: Resume trading for a previously halted symbol
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resumed
        "400":
          description: Unknown symbol
        "401":
          description: Unauthorized
        "429":
          description: Rate limited
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT